Moves other beta feature reports where they belong
The crosswiki_beta and flow_beta reports were running here because of a
limitation in reportupdater.  Reportupdater now supports multiple query
folders per repository, so we can clean up.

NOTE: all changes that share this bug should be merged / deployed~
together to prevent data loss.  Here is the list of changes for~
convenience:

https://gerrit.wikimedia.org/r/#/c/289004/
https://gerrit.wikimedia.org/r/#/c/289005/
https://gerrit.wikimedia.org/r/#/c/289006/
https://gerrit.wikimedia.org/r/#/c/289007/

Bug: T126549
Change-Id: I1c343992dbab4cfa77eb0f7c5c44fc2128669b8e